These are the main files for the website (so far)
Functionality achieved:
-Users can register
-Users Data is stored in the DB (w/ passwords hashed)
-Data is validated (server-side)
-Users are able to log in using their Details entered from the login page
-Home page for users which currently displays the $_SESSION data array and some 'Fake Data' post.
  
This website includes a register and login page which uploads and compares the entries with the real database. 
There is minimal validation for the users input but there is enough to check if the Email, Name, Password are valid (not protected against SQL injections yet). 
"IdeaSubmission" pages do not work at the moment.
When users log in, their Department, Name and Role are saved into a php Session.